Injectors (through their bindings) can dispose of injected resources.
Right now this amounts to enabling the New binding to delete a pointer.  No reference counting (etc) is done, it is still perfectly possible to dispose of something you shouldn't (such as when the dependency is consumed by someone else as well.)

Also don't do the actual delete yet, as my unit test can't prove it is required.