Create an injection cache and use it to make the test pass.
The injection cache is spread over concrete several static maps, to flexibly cater to key types.  The cache keys include a pointer to the injector itself to prevent collisions.

The New bindings insert transitive dependencies, along with a pointer to the new dependent on provision.  On disposal they first delete the value and then pass the (now garbage) pointer to the cache to transitively dispose its cached dependencies.

While it isn't well-exercised yet, Handle is a new type attempting to be a container-safe way of holding either pointers or references to a given value (and it must be constructed with either a pointer or a reference.)  It deals with references with a template specialization which simply takes the address and knows how to give a reference back.