Allow the Injector reference in Deleter to be const.
Disposal is now a const action on an injector.  I submit this is reasonable, as the disposed dependency is not part-of the injector: the injector merely knows the correct way to get rid of it.

By contrast, the NewDelete (which may turn into a proper allocator shortly) may represent a data store owning the dependency's underlying memory; delete_() should not be const.  Instead the NewDelete has been marked mutable.