Use the templated Allocator in New bindings.
This carries some implications:
 * Initializer no longer uses new and delete
 * Instead it accepts pointers as parameters and performs placement new and explicit destruction.
 * Particularly, construct()s return type is now void.

Similarly, the duties of the MockInitializer in the test suite were divided up into MockInitializer and MockAllocator.  Mocking the allocator is tricky, as one does not specify allocator instances, but rather types.  So the mock proper is MockAllocatorBacking, and an instance of this is added to the fixture class.  MockAllocatorBase then contains a static pointer to such a backing.  MockAllocator descends from MockAllocatorBase, and is templated (MockAllocatorBase is not), and satisfies enough of the allocator contract to test.