Don't dispose IfacePointers, dispose ImplPointers.
This avoids the static_cast, which near as I can tell, has no guarantee of working at all.  It also avoids feeding an IfacePointer into deallocate, which is not what allocate returned.

We pull off this monkey trick by simply holding onto the ImplPointer (in the deleter) before it gets converted to an IfacePointer for the shared_ptr.

This may be Too Clever, the jury's still out..