Switch the dependency resolution technique.
Previously we were relying on overload resolution at compile time.  Now do it by reinventing poor man's RTTI and do it at runtime.  This is less cumbersome (relieving us from having to hide c++ from itself with macros) and has a clear method for extending the definition of provision request beyond just the interface type.

The trick is a function template named BindKeyFactory.  By taking the address of instantiations, we can get opaque but comparable values that represent types (the template parameters.)  These can, for example, be put in a map as keys.  The values, called bindings, are templated derived types sharing a common untemplated base class.

To use a binding, it is first resolved into a "ResolvedBinding" that knows the return type of its provide method.  This resolution checks that the bind keys of the requested and bound types are the same, becoming our poor man's dynamic_cast.  Once we have a resolved binding, other functionality is exposed with runtime polymorphism.

Injectors are no longer templated, and modules are (currently) no longer types.  Instead modules are functions that accept a Binder as a parameter, which allows them to declare bindings.  To create an injector, one first makes a Bindings, adds one or more modules to it, and then uses a factory method on it to create the Injector.  This structure closely mirrors that of guice.