Rename BindingBuilder -> Bind.  Add Clause.
Clause is a base type for a templated fluent api.  The gag is to chain templated non-static factory methods, each producing a new intermediate representing that step of the 'grammar'.  This lets us pass both runtime information (the binding map) and template types along.  The destructors of the intermediates check a flag that beings set, but is cleared whenever a factory method is called.

Destructors for intermediates at the end of call chains therefore act, by passing the binding map to a templated action.  Note we carefully avoid calling virtual methods from the destructor.  The actions are forbidden from throwing.