Throw on incomplete bindings safely.
That is, outside of the Clause destructor.  Use what is quickly becoming my favorite trick to save a function pointer templated on our exception type in BindingMap instead (which becomes a derived class of its previous typedef map.)  Thus Clauses may throwLater<SomeException>().  The next call to bind() checks for (and throws) pending exceptions, as does Bindings::add to catch the last bind in a module.