First stab at scopes.
Cached values are kept in ScopeCache.  ScopeCache dances around the type system with a few tricks:
 * Value smart pointers are copied to heap-allocated smart pointers of the same type.
 * The copies are cast to void * wrapped in SAUCE_SHARED_PTR<void> before being inserted
 * The SAUCE_SHARED_PTR<void>s use a custom deleter that casts back and explicitly deletes.

The void * cast allows the cache to be type-agnostic.  By not unpacking the original shared pointer, but keeping a heap-allocated copy instead, we ensure the value is not disposed of out form under our feet.

Unlike guice scopes, these scopes merge the notions of entrance and exit.  Instead scopes may be cleared at any time (with a call off the injector): this is equivalent to scopes always being entered.