Make the test pass.
Hide choice of locker and lockable (if any) behind a pure virtual interface called Lock, and supply a similar pure virtual factory interface.  BaseInjectors are constructed with such a factory.  Locks are acquired at the top of Injector::get and Injector::eagerlyProvide.

This is a coarse lock, but it'll do the trick..