Extract NakedBinding from TransparentBinding.
TransparentBinding acknowledges the implementation type and deals with scoping.  However it provides unscoped instances by smart pointer in an unspecified way (pure virtual method.)

NakedBinding provides the way by providing naked pointers (the previous pure virtual), wrapping them in smart pointers and registering a custom deleter.  The custom deleter disposes of the naked pointer with the disposal method of the binding (as before.)

This will open the way to implicitly binding the injector itself, as well as ProviderBindings later.