Rename TransparentBinding -> Injection.
I'm defining an 'injection' to be the strategy used by a binding to satisfy a request for an interface.  The distinction between binding and injection is currently blurry because they stand in one-to-one relation through the class hierarchy.

However, to support setter injection (or member injection, or any other kind of injection that doesn't directly supply the Impl itself) we need to introduce a plurality.  A single binding will (soon) map not to a single injection but to a single injection and zero or more of something-else, the distinction being that Injections supply the Impl values themselves, and each something-else merely decorates or augments them.  I've almost settled on "SubInjection", but I'm not thrilled with it as a name.  InjectionDecorator or InjectionModifier are just too long.