Give libunittest merciful CPPFLAGS and LDFLAGS.
Since it's just gmock/gtest code, there's no reason to expect it should conform to the settings for the rest of the project.  Specifically, it is not -ansi compliant, but only pedantic implementations (such as cygwin) catch this.  Also there's no reason to include it in the code coverage measurements.

See http://cygwin.com/ml/cygwin/2011-10/msg00139.html