Combine provider binding declarations.
This would be straight-forward at this point, save for a hitch: Clause was initially designed to create only one binding at a time.

Particularly, subsequent calls to bind() on a ClauseState knock out previous ones.  This is a key aspect to how the fluent API works, since it allows us to eagerly call onComplete even though there may in fact be further clauses.  The binding registered in the presumptuous onComplete is simply discarded and replaced (eventually) with the true onComplete binding.

Thankfully the use case for our desired extra binding (the ProviderInjection) doesn't need any such discarding logic.  So instead of changing pendingBinding into a vector, add a vector beside it called extraPendingBindings.  This received new bindings through a completely separate method called bindExtra, though it still registers accumulated bindings in the underlying container only on destruction.

The contract exposed to clauses therefore changes.  In onComplete a clause may bind() or bindExtra().  Calls to bind() may be knocked out by subsequent clauses, but bindings registered with bindExtra() are guaranteed to survive.  Since onComplete is called by each clause (even when it is not last), clauses should therefore be careful to not make the same bindExtra() call both in onComplete and in a fluent method leading to another clause: just once in onComplete should be enough.