Add ansi_compilation_test.
gmock/gtest are not, strictly speaking, -ansi compatible.  g++ on my ubuntu & mac aren't strict enough to notice (at least, without more flags) but cygwin's environment is.  So, the unit tests are compiled without -ansi.

The only problem is that this is (as yet) a header-only library and so there are no object files left to add -ansi to, despite the library very likely being compatible.  So add ansi_compilation_test.cc, whose only purpose is to include all relevant headers and instantiated all needed templates while being compiled with -ansi.  It proves this by failing if __STRICT_ANSI__ is not defined.

Currently it does not instantiate templates, I'm just wiring it into the build.  However it did just catch a missing header I forgot.