NewInjection passes dynamic names to getHelper.
getHelper() in turn now requires a string parameter.  Only NewInjection and ProviderInjection currently use it.  I fully expect the signature of provide() to grow this parameter as well, which is where ProviderInjection will take it from.  This would preserve the "name of the provided is the name of the provider" rule.