Shove dynamic name down into ProvidingInjection.
This allows us to get at it even when only holding an injection.  Philosophically, the dynamic name should be considered part of the (normalized) dependency and this too is already available to injections.

Come to think of it, I'm not sure why bindings and injections are separate entities.  I think it's something to do with plural injections per binding for setter injection, if I ever get to it..