Extract ClauseState::finalizeInjection().
This sets the dynamic name and dynamic dependency names on the current pendingBinding.  It also clears the dynamicDependencyNames (proving they could be used elsewhere) and sets a flag.  If the flag is already set, or if no pendingBinding is defined, do nothing instead.  At the top of bind(), assert the flag has not been set.

The net effect is proving that no call to bind() will occur after the first call to finalizeInjection().

This fact will be useful later when trying to use dynamicDependencyNames for method injection.  Currently finalizeInjection() is only called in the destructor (where it was extracted from), but it will soon also be called from a bindModifier() or similar.