Add variadic appliers.
The intent here is to create a general facility for invoking functions with arbitrary signatures.  With this doing the heavy lifting, NewBinding can hopefully becoming much simpler, and creating setter injection should be similarly easier.

applyFunction(), applyMethod() and applyConstructor() accept a function pointer, method pointer or constructor type (function type) respectively, along with:
 * a type concept that supplies parameters tailor for invoking those objects
 * an arbitrary value passed to the concept type on each use.

It performs the invocation, and returns the result.

The mentioned concept is called a Parameters, and is expected to contain a member class template named Parameter like so:

    template<typename T, int i, typename Passed>
    struct Parameter {
      T yield(Passed) {
        return T();
      }
    };

The three template parameters are the expected parameter type T, the zero-based parameter index i, and the arbitrary value Passed.  The Parameter template contains the single yield() method, which accepts a Passed and does whatever is needed to produce the required T.  By using a member class template, we open the possibility of template specialization.  Consider:

    struct SpecializedParameters {
      template<typename T, int i, typename Passed>
      struct Parameter {
        T yield(Passed) {
          return T();
        }
      }

      template<int i, typename Passed>
      struct Parameter<std::string, i, Passed> {
        std::string yield(Passed) {
          return "foobar";
        }
      };
    };

Here all requested parameter types are default constructed, save for strings which return "foobar".  More interesting applications involve sending the request for a T through to an Injector.