Here's a freebee.
I suppose I should be describing this.  We're adding a Type typedef to all Parameter concepts, and using it whenever we need to materialize the concrete type of a function/method/etc in ApplyVariadic.  This will allow us to model the transform taking Named<Iface,Name> -> sauce::shared_ptr<Iface> explicitly.

We didn't need to do this before since (insanely) constructors don't really exist in C++ (to the point that you could take an address.)  So, they don't have types.  Methods *do* exist on the other hand, and we need to pass them around by address and so know their real types.