Revert sfinae for implicit modifying bindings.
This reverts the last 4 commits.

I need to think this through more before committing to a particular design, and it's not critical for the projects I just need DI for anyway (where setter injection actually will prove to be important..)

I imagine at a later date SelfBinding will show up again, along with a general case for class-specified bindings: we'll sfinae-detect 'void inject(SlightlyDifferentBinder)' or similar, allowing classes to declare (possibly restricted) bindings without a module.  With that in place I won't feel bad just responding to a few typedefs for basic implicit functionality: if the user wants to go down that route, they can.