Ditch `ansi_compilation_test` for per-standard tests.
First, `cxx98_pedantic_compilation_test` is always present and runs with `-std=c++98` and
`-pedantic-errors`.  `-ansi` is dropped since it just seems to alias a particular `-std` in
gcc and clang.

After bringing in a stock c++11 detection autoconf macro, use it to determine availability
of C++11.  If present, use it to compile `cxx11_pedantic_compilation_test`.